Backend/requirements.txt
Django>=3.2 
djangorestframework>=3.12
numpy>=1.20
pandas>=1.2
scikit-learn>=0.24
django-cors-headers==4.6.0
